==> 00-install_packages.ipynb <==
import sys
!{sys.executable} -m pip install requests pandas tensorflow==2.12.0 tqdm geopy matplotlib scikit-learn

import tensorflow as tf
print(tf.__version__)

import numpy as np
print(np.__version__)
