==> P3-scatter_plot.ipynb <==
import glacierml as gl
import matplotlib.pyplot as plt
import numpy as np
from scipy.stats import gaussian_kde
from tqdm import tqdm
from matplotlib import ticker
import configparser
import pandas as pd
import os
pd.set_option('display.max_columns', None)
import path_manager as pm
[
        home_path, data_path, RGI_path, glathida_path, 
        ref_path, coregistration_testing_path, 
        arch_test_path, LOO_path
] = pm.set_paths()

cols, df = gl.load_LOO_data(home_path,include_train = False)
# df = df.dropna(subset = ['FMT'])

x = df['FMT'] * df['Area']
y = np.mean(df[cols], axis = 1) * df['Area']
# xy = np.vstack([np.log10(x),np.log10(y)])
# print('calculating density')
# z = gaussian_kde(xy)(xy)

bins_x = np.logspace(np.log10(np.min(x)), np.log10(np.max(x)), 25)
bins_y = np.logspace(np.log10(np.min(y)), np.log10(np.max(y)), 25)

def scatter_hist(x, y, ax, 
                 ax_histx, ax_histy
                ):
    # no labels
    ax_histx.tick_params(axis="x", labelbottom=False, size = 36)
    ax_histy.tick_params(axis="y", labelleft=False, size = 36)

    # the scatter plot:
    ax.scatter(x, y, 
               # c = (z), 
               s = 100,
               cmap = 'viridis',
               marker = '.'
              )
    ax.plot(
        (0.0001,5000),
        (0.0001,5000),
        '-',
        c = 'orange',
        linewidth = 5

    )

    ax_histx.hist(
        (x), 
        bins = bins_x,
        edgecolor = 'black',
        log = True
    )

    ax_histy.hist(
        (y), 
        bins = bins_y, 
        edgecolor = 'black',
        orientation='horizontal',
        log = True
    )
    ax_histx.tick_params(axis="x", labelsize=36)
    ax_histy.tick_params(axis="y", labelsize=36)
    ax_histy.grid(alpha = 1,linewidth = 2)
    ax_histx.grid(alpha = 1,linewidth = 2)

fig = plt.figure(figsize=(20,20))
gs = fig.add_gridspec(2, 2,  width_ratios=(4, 1), height_ratios=(1, 4),
                      left=0.1, right=0.9, bottom=0.1, top=0.9,
                      wspace=0.05, hspace=0.05)


# Create the Axes.

ax = fig.add_subplot(gs[1, 0])
ax_histx = fig.add_subplot(gs[0, 0], sharex = ax)
ax_histy = fig.add_subplot(gs[1, 1], sharey = ax)


ax_histx.tick_params(axis="y", labelsize=36)
ax_histy.tick_params(axis="x", labelsize=36)

ax.tick_params(axis="both", labelsize=36)


print('Drawing scatter plot...')
scatter_hist(
    x, 
    y,
    ax, 
    ax_histx,
    ax_histy,
)
# plt.grid()
print('Drawing done')
print('Setting scales and labels...')
ax.set_yscale('log')
ax.set_xscale('log')
ax.set_ylabel('This Study Volume (km$^3$)', fontsize = 36)
ax.set_xlabel('Farinotti et al. Volume (km$^3$)', fontsize = 36)
# major_ticks = [1e-5,1e0,1e1,1e4]
# ax.set_xticks(major_ticks)
# ax.set_yticks(major_ticks)
ax.minorticks_on()

ax.grid(which = 'major',alpha = 1,linewidth = 2)
ax.grid(which = 'minor',alpha = 1,linewidth = 1,linestyle = '--')
plt.text(5e-16, 300000, 'A.',fontsize = 42)
plt.text(5e-16, 3500, 'C.',fontsize = 42)
plt.text(10, 3500, 'B.',fontsize = 42)
